refactor: divide client and server part
